Added functionality to Clear single meal
Guys. I found out that if we press the random meal button and have a random meal displayed, and then we press the search button, our random displayed meal disappears. It's really annoying. It happens because we clear a single meal in searchMeal(e) function. To avoid this problem we don't want to just clear our single.mealEl.innerHTML when we have a random meal displayed. We first what to check it like this. It works!