Something better instead of alert()
Hi Guys! So when we press the search button when we don't have any value in the input field, instead of having an alert I added something better. 
It looks hard but it's actually really easy. We check if we have something in the heading (in case if we already searched for results, we have something there), and we save the heading to var, we replace it with our message for 2 seconds, and then we return back to the heading what we had there originally. If case if we don't have anything in the heading, then we just display our message for 2 seconds ^.^